requirements.txt
Django==5.2.7
djangorestframework==3.16.1
python-decouple==3.8

# Database packages
psycopg2-binary==2.9.11  # PostgreSQL adapter
mysqlclient==2.2.6       # MySQL adapter (alternative: PyMySQL==1.1.1)
dj-database-url==2.2.0   # Database URL parsing

# Cache and Redis
redis==6.4.0
django-redis==6.0.0

# Additional packages
Pillow==12.0.0
django-cors-headers==4.6.0
django-extensions==3.2.3
coreapi==2.3.3
requests==2.32.5

# JWT Authentication
djangorestframework-simplejwt==5.3.0
